added user confirm delete

+++ src/GUI/Controllers/MainController.java
@@ -3,6 +3,7 @@ package GUI.Controllers;
 import BE.Category;
 import BE.Movie;
 import GUI.Models.MovieModel;
+import GUI.Util.ConfirmOK;
 import GUI.Util.ExceptionHandler;
 import GUI.Models.ModelsHandler;
 import GUI.Util.ModalOpener;
@@ -110,7 +111,8 @@ public class MainController extends BaseController {
     private void handleRemoveMovie(ActionEvent actionEvent) {
         Movie selectedMovie = tbvMovies.getSelectionModel().getSelectedItem();
 
-        if (selectedMovie != null) {
+        if (selectedMovie != null && ConfirmOK.confirm("Remove movie?",
+                "Are you sure you want to remove " + selectedMovie.getTitle() + "?")) {
             try {
                 getModelsHandler().getMovieModel().deleteMovie(selectedMovie);
             }
